Pulled out RigidBodyLoop into its own files.
