+++ drake/systems/plants/RigidBodyTree.h
@@ -16,6 +16,7 @@
 #include "drake/systems/plants/KinematicsCache.h"
 #include "drake/systems/plants/RigidBody.h"
 #include "drake/systems/plants/RigidBodyFrame.h"
+#include "drake/systems/plants/rigid_body_loop.h"
 #include "drake/systems/plants/collision/DrakeCollision.h"
 #include "drake/systems/plants/joints/DrakeJoint.h"
 #include "drake/systems/plants/pose_map.h"
@@ -47,27 +48,6 @@ class DRAKERBM_EXPORT RigidBodyActuator {
   const double reduction;
   const double effort_limit_min;
   const double effort_limit_max;
-};
-
-class DRAKERBM_EXPORT RigidBodyLoop {
- public:
-  //
-  // Constructs a RigidBodyLoop between two frames. Is this the correct API?
-  // TODO(amcastro-tri): review the correctness of this API
-  RigidBodyLoop(std::shared_ptr<RigidBodyFrame> _frameA,
-                std::shared_ptr<RigidBodyFrame> _frameB,
-                const Eigen::Vector3d& _axis)
-      : frameA(_frameA), frameB(_frameB), axis(_axis) {}
-
-  const std::shared_ptr<RigidBodyFrame> frameA, frameB;
-  const Eigen::Vector3d axis;
-
-  friend std::ostream& operator<<(std::ostream& os, const RigidBodyLoop& obj);
-
- public:
-#ifndef SWIG
-  EIGEN_MAKE_ALIGNED_OPERATOR_NEW
-#endif
 };
 
 class DRAKERBM_EXPORT RigidBodyTree {
